change in unittest and created /public folder

+++ unittest.sh
@@ -1,14 +1,24 @@
 #!/bin/bash
+param=$1
+loadful=${param:-"1"}
+if [[ $loadful = "1" ]]
+then
 onefile=companylist_noheader.csv
 dir1=unittest
-dir2=public
-cp $SCRIPT_DIR/datafiles/data-master.7z /tmp/
-cd /tmp
-7za x data-master.7z -o./data-master -y
+7za x $SCRIPT_DIR/datafiles/data-master.7z -o/tmp/data-master -y
 hdfs dfs -rm -r -f /$dir1
-hdfs dfs -rm -r -f /$dir2
 hdfs dfs -mkdir /$dir1
-hdfs dfs -mkdir /$dir2
 hdfs dfs -put /tmp/data-master/$onefile /$dir1
-hdfs dfs -put /tmp/data-master/* /$dir2/
 hdfs dfs -tail /$dir1/$onefile
+echo "single file loaded for unit testing"
+else
+#data load
+dir2=public
+7za x $SCRIPT_DIR/datafiles/data-master.7z -o/$dir2 -y
+hdfs dfs -rm -r -f /$dir2
+hdfs dfs -mkdir /$dir2
+hdfs dfs -put /tmp/data-master/* /$dir2/
+hdfs dfs -tail /$dir2/companylist_noheader.csv
+echo "schema public data is loaded"
+fi
+
